requirements.txt
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-oauthlib==0.4.3
requests==2.25.1